// src/enums/skill.ts
export enum Skill {
  NODE = 'Node.js',
  NEST = 'NestJS',
  TS = 'Typescript',
  MONGO = 'MongoDB',
  POSTGRES = 'PostgreSQL',
  CICD = 'CI/CD',
  AWS = 'AWS',
  AZURE = 'Azure',
  RABBIT = 'RabbitMQ',
  DOCKER = 'Docker',
  K8S = 'k8s',
  ROR = 'Ruby on Rails',
  PYTHON = 'Python',
  REACT = 'React',
  GRAPHQL = 'GraphQL',
}